Move VM disks and NICs test in their own file
